third_party/blink/web_tests/fast/events/touch/tap-highlight-color-expected.txt

Touch Me!

Check tap highlight color in touch event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Event: touch start
PASS event.targetTouches[0].target.tagName is "DIV"
PASS event.targetTouches[0].target.id is "touchtarget"
PASS div.ownerDocument.defaultView.getComputedStyle(div, null).getPropertyValue('-webkit-tap-highlight-color') is "rgba(11, 22, 33, 0)"
Event: touch end
PASS div.ownerDocument.defaultView.getComputedStyle(div, null).getPropertyValue('-webkit-tap-highlight-color') == 'rgba(11, 22, 33, 0)' is false
PASS div.ownerDocument.defaultView.getComputedStyle(div, null).getPropertyValue('-webkit-tap-highlight-color') is non-null.
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
